Move model/training classes within gpsr namespace
